알림 기능

+++ src/main/java/com/example/demo/repository/ChecklistRepository.java
@@ -4,5 +4,5 @@ import com.example.demo.domain.Checklist;
 import com.example.demo.domain.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ChecklistRepository extends JpaRepository<Checklist, Long> {
+public interface ChecklistRepository extends JpaRepository<Checklist, Integer> {
 }
